// src/templates/formEditarAvalCaderno/types/FormData.ts
export interface FormData {
  nomeAvaliado: string;
  matricula: number;
  id: number;
  conteudosRelevantes: string;
  questao1: string;
  questao2: string;
  questao3: string;
  questao4: string;
  questao5: string;
  questao6: string;
  questao7: string;
  questao8: string;
  questao9: string;
  questao10: string;
  conclusaoAvaliador: string;
}